Started Integrating Redux & Firebase
Adding redux store and firebase keys to start with login. After that,
managin each users collections of expenses will be done.
