+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "expo": "^32.0.0",
+    "firebase": "^5.9.0",
     "react": "16.5.0",
     "react-native": "0.57.1",
     "react-native-action-button": "^2.8.5",
@@ -17,7 +18,9 @@
     "react-native-vector-icons": "^6.4.1",
     "react-navigation": "^3.3.2",
     "react-redux": "^6.0.1",
+    "react-redux-firebase": "^2.2.6",
     "redux": "^4.0.1",
+    "redux-firestore": "^0.7.2",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
